requirements.txt
numpy==1.19.1
streamlit==1.2.0
pandas==1.0.5
keras==2.7.0
Keras-Preprocessing==1.1.2
tensorflow-cpu==2.7.0
matplotlib==3.3.4
pillow==6.2.0